pyproject.toml
[tool.poetry]
name = "goit-web-hw-10"
version = "0.1.0"
description = "Homework GoIT Python web"
package-mode = false
authors = ["Yevhenii Baranov <j52b10i78@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
mongoengine = "^0.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
